+++ xmltemplate/yarp/yarpopenpolydrive.xml
@@ -1,0 +1,1 @@
+<command library="yarp" name="yarpopenpolydrive" polydrivertag="" device="" repetitions="1" wait="0" reporterror="true"></command>
